+++ activity/order_filler_test.go
@@ -30,7 +30,7 @@ func TestCanFillBuyOrder(t *testing.T) {
 
 	symbolService := SymbolServiceStub{}
 
-	amount := float32(1000.00)
+	amount := float64(1000.00)
 	symbol := "aapl"
 	o, err := orderRepo.Save(order.NewBuyOrder(
 		account.ID,
